pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "spearhead"
version = "1.0.0"
description = "A production-grade, offline-first battalion readiness system."
authors = [
    {name = "Principal Software Architect", email = "architect@spearhead.com"},
]
requires-python = ">=3.9"
dependencies = [
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pandas>=2.0.0",
    "jinja2>=3.0.0",
    "typer>=0.9.0",
    "pyyaml>=6.0",
    "fastapi>=0.110.0",
    "google-auth>=2.35.0",
    "requests>=2.32.3",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
]

[project.scripts]
spearhead = "spearhead.main:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
